+++ javacutil/src/main/java/org/checkerframework/javacutil/AnnotationMirrorSet.java
@@ -44,6 +44,16 @@ public class AnnotationMirrorSet
   /** Default constructor. */
   public AnnotationMirrorSet() {}
 
+  // TODO: Should this be an unmodifiable set?
+  /**
+   * Creates a new {@link AnnotationMirrorSet} that contains {@code value}.
+   *
+   * @param value the AnnotationMirror to put in the set
+   */
+  public AnnotationMirrorSet(AnnotationMirror value) {
+    this.add(value);
+  }
+
   /**
    * Returns a new {@link AnnotationMirrorSet} that contains the given annotation mirrors.
    *