+++ docs/CHANGELOG.md
@@ -62,6 +62,12 @@ Signedness Checker:
 
 **Implementation details:**
 
+Methods that now return `AnnotationMirrorSet` instead of `Set<? extends AnnotationMirror>`:
+ * `getTopAnnotations()`
+ * `getBottomAnnotations()`
+ * `getDefaultTypeDeclarationBounds()`
+ * `getExceptionParameterLowerBoundAnnotations()`
+
 **Closed issues:**